requirements.txt
Django==4.2.14
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
gunicorn==23.0.0
psycopg2-binary==2.9.10
pytest==8.3.4
pytest-django==4.9.0
pytest-mock==3.14.0
requests==2.32.3
model-bakery==1.20.0
pytest-responses==0.5.1


ipython==8.30.0
